Partial #10, #7, #19. Closes #4.
Watercourses are now added and named from OS Open Rivers. Widths
calculated against the geometries which exist in the surface
layer. Output is now moved to the product directory when generated
with gm-tile, which now also checks if terrain data exists and
attempts to produce including downloading LiDAR. NRW not currently
included.